% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.ratio.plot.R
\name{control.ratio.plot}
\alias{control.ratio.plot}
\title{A Function for auto-generating control ratio plots.
Designed for data outputed from soft.max.clean function}
\usage{
control.ratio.plot(data, positive.control)
}
\arguments{
\item{data}{Data formatted per soft.max.clean output.}

\item{positive.control}{Name of positive control, as string.}
}
\description{
A Function for auto-generating control ratio plots.
Designed for data outputed from soft.max.clean function
}
\examples{
control.ratio.plot()
}
