% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.max.clean.R
\name{soft.max.clean}
\alias{soft.max.clean}
\title{A Function for formatting BKA excel data}
\usage{
soft.max.clean(row_start_of_data, file_path, num_of_time_points)
}
\arguments{
\item{row_start_of_data}{Row number of start of data in Excel.}

\item{file_path}{Where does the data live on your computer?}

\item{num_of_time_points}{Number of time points.}
}
\description{
A Function for formatting BKA excel data
}
\examples{
soft.max.clean
}
